decoupling turn policy
