+++ ga.py
@@ -573,4 +573,16 @@ class GenericAgentHandler(BaseHandler):
         for hook in getattr(self.parent, '_turn_end_hooks', {}).values(): hook(locals())  # current readonly
         return next_prompt
 
-# get_global_memory() moved to turn_policy.py
+def get_global_memory():
+    """读取全局记忆并格式化"""
+    prompt = "\n"
+    try:
+        suffix = '_en' if os.environ.get('GA_LANG', '') == 'en' else ''
+        with open(os.path.join(script_dir, 'memory/global_mem_insight.txt'), 'r', encoding='utf-8', errors='replace') as f: insight = f.read()
+        with open(os.path.join(script_dir, f'assets/insight_fixed_structure{suffix}.txt'), 'r', encoding='utf-8') as f: structure = f.read()
+        prompt += f'cwd = {os.path.join(script_dir, "temp")} (./)\n'
+        prompt += "\n[Memory] (../memory)\n"
+        prompt += structure + '\n../memory/global_mem_insight.txt:\n'
+        prompt += insight + "\n"
+    except FileNotFoundError: pass
+    return prompt
